+++ src/lib/pages/public-project/components/DetailHeader.tsx
@@ -11,6 +11,7 @@ import { MdChevronRight } from "react-icons/md";
 
 import { AppLink } from "lib/components/AppLink";
 import type { Option, PublicDetail } from "lib/types";
+import { getNameAndDescriptionDefault } from "lib/utils";
 
 import { BookmarkButton } from "./BookmarkButton";
 import { SocialMedia } from "./SocialMedia";
@@ -56,19 +57,21 @@ export const DetailHeader = ({ details, slug }: DetailHeaderProps) => {
       >
         <Box>
           <Flex gap={2} align="center">
-            <Image
-              src={details?.logo}
-              borderRadius="full"
-              alt="Celatone"
-              width={8}
-              height={8}
-            />
+            {details?.logo && (
+              <Image
+                src={details?.logo}
+                borderRadius="full"
+                alt="Celatone"
+                width={8}
+                height={8}
+              />
+            )}
             <Heading as="h5" variant="h5" className="ellipsis">
-              {details?.name}
+              {getNameAndDescriptionDefault(details?.name)}
             </Heading>
           </Flex>
           <Text variant="body2" color="text.dark" mt={2}>
-            {details?.description}
+            {getNameAndDescriptionDefault(details?.description)}
           </Text>
         </Box>
         <Flex alignItems="center" gap={4}>